Adds CLI Support
Via `clap`.
