+++ Cargo.toml
@@ -10,12 +10,15 @@ keywords = ["config", "web", "service", "remote"]
 categories = ["development-tools", "web-programming", "config", "database"]
 readme = "README.md"
 edition = "2021"
+publish = false
 
 [dependencies]
-anyhow = "1.0.57"
-axum = "0.5.4"
+anyhow = { version = "1.0.56", features = ["backtrace"] }
+axum = "0.5.9"
 home = "0.5.3"
+clap = { version = "3.2.6", features = ["derive", "env"] }
+tracing = "0.1.32"
 serde = { version = "1.0.137", features = ["derive"] }
 serde_json = "1.0.81"
 tempfile = "3.3.0"
-tokio = { version = "1.18.1", features = ["full"] }
+tokio = { version = "^1", features = ["full"] }
