+++ src/main.rs
@@ -1,10 +1,17 @@
-use anyhow::Result;
+use std::process;
 
+use cadre::cli::Cli;
+use clap::Parser;
+use tracing::error;
+
+/// Main entry point for the `cadre` binary.
 #[tokio::main]
-async fn main() -> Result<()> {
-    let app = cadre::server().await?;
-    axum::Server::bind(&"0.0.0.0:3000".parse()?)
-        .serve(app.into_make_service())
-        .await?;
-    Ok(())
+async fn main() {
+    match Cli::parse().run().await {
+        Ok(()) => process::exit(0),
+        Err(err) => {
+            error!("{err:?}");
+            process::exit(1)
+        }
+    }
 }
